extract file selection logic for easier testing
